Refactor: Dialogs and add dialogs
