+++ example/lib/example5_themes_playground/widgets/dialogs/import_settings_dialog.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+/// Dialog to confirm if user wants to import a Playground setup.
+class ImportSettingsDialog extends StatelessWidget {
+  const ImportSettingsDialog({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: const Text('Import Playground Settings'),
+      content: const Text('Import the settings from the staging area?\n'
+          '\n'
+          'You current configuration will be overwritten, and replaced\n'
+          'with the settings from the staging area.'),
+      actions: <Widget>[
+        TextButton(
+            onPressed: () {
+              Navigator.of(context).pop(false);
+            },
+            child: const Text('CANCEL')),
+        TextButton(
+            onPressed: () {
+              Navigator.of(context).pop(true);
+            },
+            child: const Text('IMPORT')),
+      ],
+    );
+  }
+}
